+++ activerecord/test/cases/associations/has_many_through_associations_test.rb
@@ -1165,4 +1165,11 @@ class HasManyThroughAssociationsTest < ActiveRecord::TestCase
 
     assert_nil Club.new.special_favourites.distinct_value
   end
+
+  def test_association_force_reload_with_only_true_is_deprecated
+    post = Post.find(1)
+
+    assert_deprecated(/reload: true/) { post.people(true) }
+    assert_not_deprecated { post.people(reload: true) }
+  end
 end