+++ activerecord/test/cases/associations/has_one_associations_test.rb
@@ -607,4 +607,11 @@ class HasOneAssociationsTest < ActiveRecord::TestCase
       end
     end
   end
+
+  def test_association_force_reload_with_only_true_is_deprecated
+    firm = Firm.find(1)
+
+    assert_deprecated(/reload: true/) { firm.account(true) }
+    assert_not_deprecated { firm.account(reload: true) }
+  end
 end